indexer implementation, adjust the point scheme, update dependencies and java
