+++ Searcher/src/main/java/com/github/pathfinder/service/IPointConnector.java
@@ -1,0 +1,9 @@
+package com.github.pathfinder.service;
+
+import java.util.List;
+
+public interface IPointConnector {
+
+    void createConnections(List<Integer> chunkIds);
+
+}
