+++ Searcher/src/main/java/com/github/pathfinder/service/IPointService.java
@@ -7,6 +7,4 @@ public interface IPointService {
 
     List<PointNode> saveAll(int chunkId, List<PointNode> nodes);
 
-    void createConnections(List<Integer> chunkIds);
-
 }
