+++ Searcher/src/test/java/com/github/pathfinder/configuration/SearcherNeo4jTest.java
@@ -1,7 +1,6 @@
 package com.github.pathfinder.configuration;
 
 import com.github.pathfinder.service.impl.ChunkService;
-import com.github.pathfinder.service.impl.PointConnector;
 import com.github.pathfinder.service.impl.PointService;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -24,7 +23,6 @@ import static org.springframework.transaction.annotation.Propagation.NOT_SUPPORT
 @Import({Neo4jTestConfiguration.class,
         CoordinateConfiguration.class,
         Neo4jTestTemplate.class,
-        PointConnector.class,
         PointService.class,
         ChunkService.class})
 public @interface SearcherNeo4jTest {
