+++ docker-compose.yml
@@ -20,6 +20,9 @@ services:
       PORT: ${SECURITY_PORT}
     ports:
       - "${SECURITY_PORT}:${SECURITY_PORT}"
+    depends_on:
+      postgresql:
+        condition: service_healthy
   indexer:
     build: ./Indexer
     container_name: indexer
@@ -28,6 +31,9 @@ services:
       PORT: ${INDEXER_PORT}
     ports:
       - "${INDEXER_PORT}:${INDEXER_PORT}"
+    depends_on:
+      postgresql:
+        condition: service_healthy
   searcher:
     build: ./Searcher
     container_name: searcher
@@ -61,10 +67,11 @@ services:
     container_name: neo4j
     environment:
       NEO4J_AUTH: ${NEO4J_USERNAME}/${NEO4J_PASSWORD}
-      NEO4J_dbms_default__listen__address: neo4j
+      NEO4J_server_default__listen__address: neo4j
       NEO4J_PLUGINS: '["apoc"]'
-      NEO4J_dbms_default__advertised__address: neo4j
-      NEO4J_dbms_connector_bolt_advertised__address: :${NEO4J_PORT}
+      NEO4J_dbms_memory_transaction_total_max: 0
+      NEO4J_server_default__advertised__address: neo4j
+      NEO4J_server_bolt_advertised__address: :${NEO4J_PORT}
       NEO4J_initial_dbms_default__database: ${NEO4J_DB_NAME}
     restart: on-failure
   liquibase:
